requirements.txt
pytest>=8.3.0
PyYAML>=6.0
